+++ internal/mcp/actor/actor.go
@@ -1,9 +1,9 @@
 package actor
 
 type Actor struct {
-	UserID     uint     `json:"user_id"`
-	TenantID   uint     `json:"tenant_id"`
-	LocationID uint     `json:"location_id"`
+	UserID     uint64   `json:"user_id"`
+	TenantID   uint64   `json:"tenant_id"`
+	LocationID uint64   `json:"location_id"`
 	Email      string   `json:"email"`
 	Type       string   `json:"type"`
 	Roles      []string `json:"roles"`
