+++ internal/models/dto/rbac.go
@@ -1,7 +1,7 @@
 package dto
 
 type AddRole struct {
-	TenantID    uint     `json:"TenantID" binding:"required"`
+	TenantID    uint64   `json:"TenantID" binding:"required"`
 	Name        string   `json:"Name" binding:"required"`
 	Permissions []string `json:"Permissions" binding:"required"`
 }
@@ -12,6 +12,6 @@ type UpdateRole struct {
 	Action      int      `json:"Action" binding:"oneof=0 1"`
 }
 type UpdateRoles struct {
-	TenantID uint         `json:"TenantID" binding:"required"`
+	TenantID uint64       `json:"TenantID" binding:"required"`
 	Data     []UpdateRole `json:"Data" binding:"required"`
 }
