+++ internal/models/entity/wallet.go
@@ -3,19 +3,19 @@ package entity
 import "time"
 
 type WalletBalance struct {
-	ID       uint  `gorm:"primaryKey;autoIncrement"`
-	TenantID uint  `gorm:"not null;index:idx_tx_wallet"`
-	UserID   uint  `gorm:"not null;index:idx_tx_wallet"`
-	Balance  int64 `gorm:"not null"`
+	ID       uint64 `gorm:"primaryKey;autoIncrement"`
+	TenantID uint64 `gorm:"not null;index:idx_tx_wallet"`
+	UserID   uint64 `gorm:"not null;index:idx_tx_wallet"`
+	Balance  int64  `gorm:"not null"`
 
 	UpdatedAt time.Time `gorm:"precision:6"`
 }
 
 type WalletTransaction struct {
-	ID uint `gorm:"primaryKey;autoIncrement"`
+	ID uint64 `gorm:"primaryKey;autoIncrement"`
 
-	TenantID      uint   `gorm:"not null;index:idx_tx_wallet"`
-	UserID        uint   `gorm:"not null;index:idx_tx_wallet"`
+	TenantID      uint64 `gorm:"not null;index:idx_tx_wallet"`
+	UserID        uint64 `gorm:"not null;index:idx_tx_wallet"`
 	Amount        int64  `gorm:"not null"`
 	Type          string `gorm:"type:varchar(16);not null"` // credit, debit
 	BalanceAfter  int64  `gorm:"not null"`
@@ -32,16 +32,16 @@ func (WalletTransaction) TableName() string {
 }
 
 type WalletRecharge struct {
-	ID                  uint   `gorm:"primaryKey;autoIncrement"`
-	TenantID            uint   `gorm:"not null;index"`
-	UserID              uint   `gorm:"not null;index"`
-	Amount              int64  `gorm:"not null"`
-	PaymentGateway      string `gorm:"type:varchar(64);not null"`  // razorpay
-	OrderIDReceipt      string `gorm:"type:varchar(128);not null"` // razorpay order id receipt
-	OrderID             string `gorm:"type:varchar(128);not null"` // razorpay order id
-	Status              string `gorm:"type:varchar(32);not null"`  // pending, success, failed
-	PaymentID           string `gorm:"type:varchar(128);not null"` // razorpay payment id
-	WalletTransactionID *uint  `gorm:"index"`
+	ID                  uint64  `gorm:"primaryKey;autoIncrement"`
+	TenantID            uint64  `gorm:"not null;index"`
+	UserID              uint64  `gorm:"not null;index"`
+	Amount              int64   `gorm:"not null"`
+	PaymentGateway      string  `gorm:"type:varchar(64);not null"`  // razorpay
+	OrderIDReceipt      string  `gorm:"type:varchar(128);not null"` // razorpay order id receipt
+	OrderID             string  `gorm:"type:varchar(128);not null"` // razorpay order id
+	Status              string  `gorm:"type:varchar(32);not null"`  // pending, success, failed
+	PaymentID           string  `gorm:"type:varchar(128);not null"` // razorpay payment id
+	WalletTransactionID *uint64 `gorm:"index"`
 
 	CreatedAt time.Time `gorm:"precision:6"`
 	UpdatedAt time.Time `gorm:"precision:6"`
